+++ src/menai/menai_compiler.py
@@ -20,6 +20,7 @@ from menai.cfg.menai_cfg_optimization_pass import MenaiCFGOptimizationPass
 from menai.cfg.menai_cfg_branch_const_prop import MenaiCFGBranchConstProp
 from menai.cfg.menai_cfg_simplify_blocks import MenaiCFGSimplifyBlocks
 from menai.cfg.menai_cfg_collapse_phi_chains import MenaiCFGCollapsePhiChains
+from menai.cfg.menai_cfg_dead_captures import MenaiCFGDeadCaptures
 from menai.cfg.menai_cfg_type_propagation import MenaiCFGTypePropagation
 from menai.vcode.menai_vcode_builder import MenaiVCodeBuilder
 from menai.ir.menai_ir_builder import MenaiIRBuilder
@@ -70,6 +71,7 @@ class MenaiCompiler:
             MenaiCFGBranchConstProp(),
             MenaiCFGSimplifyBlocks(),
             MenaiCFGTypePropagation(),
+            MenaiCFGDeadCaptures(),
         ]
         self.vcode_builder = MenaiVCodeBuilder()
         self.bytecode_builder = MenaiBytecodeBuilder()